forecast_weather/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
environs==9.5.0
executing==0.8.3
idna==3.3
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib-inline==0.1.3
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
